Nexus threads prematurely terminating sessions
There exists a neutron process with both a Nexus replay thread and
a config thread.  These threads share the same ssh session.  When
they are talking to Nexus at the same time, I saw one thread closing
the nexus ssh session before the other thread was complete resulting
in unexpected errors.

This change set places a lock around the get and edit operations (and
other sensitive operational variable).  The get and edit operations
encompass the connect & close calls.  So within a single
acquire/release of the driver lock, it will contain the following
  - connect
  - edit or get
  - close

A single thread will complete these operations before proceeding
to the next thread.  There are only 2 threads contending for the
drivers which are the replay thread and config thread (which only
one config is running at the time).

Change-Id: I236504b9347b00a76c3190adbd6b0eb2c3fe3b1c
Closes-bug:  #1608747