Allow routers to be seen when changing L3 plugin
If a neutron router was created by a different L3 plugin, when
the Cisco Router Service Plugin is selected and neutron restarted,
then the pre-existing neutron router will not be found when queried
because it doesn't have a RouterHostingDeviceBinding. This patch
adds the namespace binding for routers that don't have one, which
allows them to be seen when queried.

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>