Add fix for unrouteable subnets
With GBP workflow, the unrouteable subnet for the external
segment should not be used as the primary IP on the ASR's external
VLAN sub-interface (i.e. the one that connects to the outside world).

This needs to be fixed, so that unrouteable subnets are not used
for configuring the primary IP. This patch changes the configuration
agent driver to use a subnet from the extra_subnets field instead of
the unrouteable subnet.

Closes-issue: 147