Fix upstream gate jobs
The networking-cisco project has been unmaintained, with gate jobs
failing due to bit rot. In order to get back to passing gate jobs,
several patches will be introduced to bring the code in this
repository up to stable/2023.1 (Antelope), so that subsequent
patches can enable support for the current release (stable/2024.2)
with backports for two prior releases, as per upstream policy.

This patch ensures that the unit tests are run, but marks the unit
test gate jobs as non-voting, since the patch does not ensure that the
tests will pass. Decisions will need to be made about what code to
keep and what code to deprecate in the repo, based on current use cases.

Change-Id: I406fd61097f3defd2be8ef7d76815f724d996608