Add LLDP Topology Agent
This adds a generic LLDP agent, as well as switch-specific handlers
that convert the raw topology information into tuples that can be
sent using RPCs to the server.

The agent can be configured with multiple handlers, and each handler
returns a dictionary of topology peers that they recognize, keyed by
the name of the local interface where they were seen.

A place holder exists for implementing the server-side handling,
which can be addressed in a separate patch.

Change-Id: Ie6374085a309d64226af7d80e6aa0877081c3708